Move ImageSlide and ChakraCarousel to components
