+++ client/src/components/ChakraCarousel.tsx
